Wait/test Any returns SIZE_MAX if no completions
Also, some minor example code fixes

+++ example_code/shmem_wait_until_any_all2all_sum.c
@@ -28,7 +28,7 @@ int main(void)
   for (int i = 0; i < npes; i++)
       shmem_p(&flags[mype], 1, i);
   
-  int completed_idx;
+  size_t completed_idx;
   for (int i = 0; i < npes; i++) {
       completed_idx = shmem_wait_until_any(flags, npes, status, SHMEM_CMP_NE, 0);
       for (int j = 0; j < N; j++)
